add exchangeable currency

+++ app/controllers/exchangeable_currency_controller.rb
@@ -1,5 +1,6 @@
 class ExchangeableCurrencyController < ApplicationController
   skip_before_action :verify_authenticity_token
+
   def index
     unless params[:from].present?
       currency = ExchangeableCurrency.all
@@ -10,7 +11,7 @@ class ExchangeableCurrencyController < ApplicationController
   end
 
   def show
-    render json: current_currency, status: :ok
+    render json: ExchangeableCurrencyService.new(params[:id]).currency_found?
   end
 
   def create
@@ -18,24 +19,28 @@ class ExchangeableCurrencyController < ApplicationController
   end
 
   def destroy
-    currency = current_currency
-    currency.destroy
-    render json: { status: 200 }
+    currency = ExchangeableCurrencyService.new(params[:id]).current_currency
+    if currency.present?
+      currency.destroy
+      render json: { status: 200 }
+    else
+      ExchangeableCurrencyService.not_found
+    end
   end
 
   def update
-    currency = current_currency
-    currency.update_attributes(currency_params)
-    render json: current_currency, status: :ok
+    currency = ExchangeableCurrencyService.new(params[:id]).current_currency
+    if currency.present?
+      currency.update_attributes(currency_params)
+      render json: currency, status: :ok
+    else
+      ExchangeableCurrencyService.not_found
+    end
   end
 
   private
   def currency_params
     params.require(:exchangeable_currency).permit(:from, :to)
-  end
-
-  def current_currency
-    ExchangeableCurrency.find(params[:id])
   end
 
   def filtering_params
